Split classes out and removed blank line per codefactor policy.
