+++ Blish HUD/Common/UI/Presenters/IControlPresenter.cs
@@ -1,15 +1,8 @@
-﻿using Blish_HUD.Common.UI.Views;
-using Blish_HUD.Controls;
-using Blish_HUD.Graphics.UI;
+﻿using Blish_HUD.Graphics.UI;
 
 namespace Blish_HUD.Common.UI.Presenters {
 
     public interface IControlPresenter : IPresenter {
-
-    }
-
-    public interface IControlPresenter<TControl> : IPresenter<ControlView<TControl>>, IControlPresenter
-        where TControl : Control {
 
     }
